jooq plus fixes

+++ project/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/domain/service/LinkServiceImpl.java
@@ -1,12 +1,12 @@
-package ru.tinkoff.edu.java.scrapper.domain.jdbc;
+package ru.tinkoff.edu.java.scrapper.domain.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import parser.Parser;
 import results.ParsingResult;
 import results.StackOverflowParsingResult;
 import ru.tinkoff.edu.java.scrapper.domain.exception.LinkNotFoundException;
-import ru.tinkoff.edu.java.scrapper.domain.repository.JdbcLinkRepository;
-import ru.tinkoff.edu.java.scrapper.domain.service.LinkService;
+import ru.tinkoff.edu.java.scrapper.domain.repository.LinkRepository;
 import ru.tinkoff.edu.java.scrapper.dto.domain.Link;
 import ru.tinkoff.edu.java.scrapper.dto.webclient.StackOverflowQuestionResponse;
 import ru.tinkoff.edu.java.scrapper.web.webclient.client.StackOverflowClient;
@@ -15,30 +15,33 @@ import java.net.URI;
 import java.util.List;
 
 @Component
-public class JdbcLinkService implements LinkService {
+@RequiredArgsConstructor
+public class LinkServiceImpl implements LinkService {
 
-    private final JdbcLinkRepository linkRepository;
+    private final LinkRepository linkRepository;
     private final StackOverflowClient stackOverflowClient;
     private final Parser parser = new Parser();
-
-    public JdbcLinkService(JdbcLinkRepository linkRepository, StackOverflowClient stackOverflowClient) {
-        this.linkRepository = linkRepository;
-        this.stackOverflowClient = stackOverflowClient;
-    }
 
     @Override
     public Link add(long chatId, URI url) {
         ParsingResult parsingResult = parser.parse(url.toString());
         if (parsingResult instanceof StackOverflowParsingResult stackOverflowParsingResult) {
             StackOverflowQuestionResponse response = stackOverflowClient.fetchQuestion(stackOverflowParsingResult.id());
-            return linkRepository.addAnswerComment(chatId, url, response.answerCount(), response.commentCount());
+            linkRepository.addAnswerComment(chatId, url, response.answerCount(), response.commentCount());
+        } else {
+            linkRepository.add(chatId, url);
         }
-        return linkRepository.add(chatId, url);
+        return linkRepository.getLinksByUrl(url).get(0);
     }
 
     @Override
     public Link remove(long chatId, URI url) throws LinkNotFoundException {
-        return linkRepository.remove(chatId, url);
+        List<Link> links = linkRepository.getLinksByUrl(url);
+        if (links.size() == 0) {
+            throw new LinkNotFoundException("Link not found: " + url.toString());
+        }
+        linkRepository.remove(chatId, url);
+        return links.get(0);
     }
 
     @Override
